backend/requirements.txt
# Core framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Audio processing
openai==1.12.0
pydub==0.25.1

# LLM integration
anthropic==0.18.1
groq==0.4.2

# External service integrations
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
caldav==1.3.9
vobject==0.9.6.1

# Database & storage
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
greenlet==3.0.3

# Utilities
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
boto3==1.34.34

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Production
gunicorn==21.2.0
mangum==0.17.0
